Refactor: Extract parameter parsing to dedicated module
Move parameter parsing code from the large parser.rs file to a dedicated
parameters.rs module to improve code organization and maintainability.

Changes:
- Created new src/txxt_nano/parser/parameters.rs module
- Moved ParameterWithSpans struct to parameters module
- Moved convert_parameter function to parameters module
- Moved parse_parameters_from_tokens function to parameters module
- Moved 4 parameter-related unit tests to parameters module
- Updated module declarations and imports
- All 88 lib tests + 13 parameter tests + 21 lexer tests pass

This refactoring reduces parser.rs from ~2600 lines to ~2475 lines
while maintaining all functionality and test coverage.
